requirements-dev.txt
pyqt6-tools
opencv-python
vidgear
pyscreenshot
loguru
pynput